j'ai change la position du player en double et maintenant ca avance

+++ includes/structs.h
@@ -6,7 +6,7 @@
 /*   By: skuor <skuor@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/12/17 12:48:34 by skuor             #+#    #+#             */
-/*   Updated: 2026/01/20 17:16:39 by skuor            ###   ########.fr       */
+/*   Updated: 2026/01/21 10:26:38 by skuor            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -39,8 +39,8 @@ typedef struct s_mlx
 
 typedef struct s_player
 {
-	int		pos_row;
-	int		pos_col;
+	double		pos_row;
+	double		pos_col;
 	char	spawn_dir;
 }			t_player;
